requirements.txt
google-generativeai
langchain-google-genai
langchain
streamlit
pyperclip
huggingface_hub
python-dotenv
SpeechRecognition
PyAudio
pdfplumber
python-docx
pylatexenc